requirements.txt
audioop-lts==0.2.1  ; python_version >= '3.13'
Flask==3.1.1
Flask-Cors==6.0.1
gunicorn==23.0.0
git+https://github.com/jplexer/pyspeex.git
vosk==0.3.45
pyspeex-noise==1.0.2